# docs/openapi.yaml
---
openapi: 3.0.0
info:
  title: AR.IO Observer
  version: 0.0.1
  description: |
    AR.IO Observer
components:
  schemas:
    ArweaveAddress:
      type: string
    ArweaveWallet:
      type: string
      nullable: true
    ArweaveId:
      type: string
      nullable: true
    DataHash:
      type: string
      nullable: true
    Timestamp:
      type: integer
    Evaluation:
      type: boolean
    HealthCheck:
     type: object
     properties:
       uptime: { type: number }
       date: { 
        type: string,
        format: date
       }
       message: { type: string }
    OwnershipAssessment:
      type: object
      properties:
        expectedWallet: { '$ref': '#/components/schemas/ArweaveWallet' } 
        observedWallet: { '$ref': '#/components/schemas/ArweaveWallet' } 
        failureReason: { type: string }
        pass: { '$ref': '#/components/schemas/Evaluation' } 
    ArnsAssessment:
      type: object
      properties:
        assessedAt: { '$ref': '#/components/schemas/Timestamp'}
        expectedId: { '$ref': '#/components/schemas/ArweaveId' }
        resolvedId: { '$ref': '#/components/schemas/ArweaveId' }
        expectedDataHash: { '$ref': '#/components/schemas/ArweaveId' }
        resolvedDataHash: { '$ref': '#/components/schemas/ArweaveId' }
        pass: { '$ref': '#/components/schemas/Evaluation' }
      required:
        - assessedAt
        - expectedId
        - resolvedId
        - expectedDataHash
        - resolvedDataHash
        - pass
    GatewayArnsAssessment:
      type: object
      properties:
        prescribedNames:
          type: object
          additionalProperties: { '$ref': '#/components/schemas/ArnsAssessment' }
        chosenNames:
          type: object
          additionalProperties: { '$ref': '#/components/schemas/ArnsAssessment' }
      required:
        - prescribedNames
        - chosenNames
    GatewayAssessment:
      type: object
      properties:
        ownershipAssessment: { '$ref': '#/components/schemas/OwnershipAssessment' }
        arnsAssessments: { '$ref': '#/components/schemas/GatewayArnsAssessment' }
        pass: { '$ref': '#/components/schemas/Evaluation' }
    GatewayAssessments:
      type: object
      additionalProperties: { '$ref': '#/components/schemas/GatewayAssessment' }
    Report:
      type: object
      properties:
        observerAddress: { '$ref': '#/components/schemas/ArweaveAddress' }
        generatedAt: { '$ref': '#/components/schemas/Timestamp' } 
        gatewayAssessments: { '$ref': '#/components/schemas/GatewayAssessments' }
      required:
        - observerAddress
        - generatedAt
        - gatewayAssessments
    PendingReport:
      type: object
      properties:
        message: 
          type: string
      required:
        - message

paths:
  '/ar-io/observer/healthcheck':
    get:
      responses:
        '200':
          description: |-
            200 response
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/HealthCheck'
  '/ar-io/observer/reports/current':
    get:
      responses:
        '200':
          description: |-
            200 response
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/Report'
        '202':
          description: |-
            202 response
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/PendingReport'